refactor a bit; fix logic to utilize root fields in infojson
